+++ src/definitions/users_definitions.rs
@@ -1,9 +1,9 @@
 use super::ranks::{Ranks, STRanks, SableRanks};
 use serde::{Deserialize, Serialize};
 
-pub static WIJ_ID: u64 = 3747606;
-pub static ST_ID: u64 = 3758883;
-pub static SABLE_ID: u64 = 5430057;
+pub static WIJ_ID: u32 = 3747606;
+pub static ST_ID: u32 = 3758883;
+pub static SABLE_ID: u32 = 5430057;
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct Divisions {
@@ -16,16 +16,16 @@ pub struct Divisions {
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct User {
     #[serde(default)]
-    pub user_id: u64,
+    pub user_id: u32,
     pub name: String,
     #[serde(default)]
-    pub points: u64,
+    pub points: u32,
     #[serde(default)]
-    pub events: u64,
+    pub events: u32,
     #[serde(default)]
-    pub floor_points: Option<u64>,
+    pub floor_points: Option<u32>,
     #[serde(default)]
-    pub goal_points: Option<u64>,
+    pub goal_points: Option<u32>,
     #[serde(default)]
     pub rank: Ranks,
     pub divisions: Option<Divisions>,
@@ -34,16 +34,16 @@ pub struct User {
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct DeserializeUser {
     #[serde(default)]
-    pub user_id: u64,
+    pub user_id: u32,
     pub name: String,
     #[serde(default)]
-    pub points: u64,
+    pub points: u32,
     #[serde(default)]
-    pub events: u64,
+    pub events: u32,
     #[serde(default)]
-    pub floor_points: Option<u64>,
+    pub floor_points: Option<u32>,
     #[serde(default)]
-    pub goal_points: Option<u64>,
+    pub goal_points: Option<u32>,
     #[serde(default)]
     pub rank: String,
     pub divisions: Option<Divisions>,
